ai-trading-poc/ai/requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
lightgbm==4.0.0
requests==2.31.0
python-dotenv==1.0.0
pyyaml==6.0
scipy==1.11.1
